fix: fix type mismatch in ReadingStatsStore
Convert Int to Long for readingTimeMs field to fix compilation error.

Bump versionCode 38→39, versionName 3.6.1→3.6.2 (patch)

+++ app/src/main/kotlin/com/exapps/mangaworld/core/data/ReadingStatsStore.kt
@@ -109,7 +109,7 @@ class ReadingStatsStore @Inject constructor(
             DailyStat(
                 date = date,
                 pagesRead = pagesMap[date] ?: 0,
-                readingTimeMs = timeMap[date] ?: 0L
+                readingTimeMs = (timeMap[date] ?: 0).toLong()
             )
         }
     }
